WIP: Migrate to hilt
